refactor(in-memory): split checkout session store
- move checkout-session selection and store operations into checkout-session-store
- keep the repository Layer file focused on service construction
- preserve repository behavior while reducing the Fallow hotspot on the public module
